Added a Tag Selector attribute
Custom property drawer to access tag enums from a list.

From a script, use [TagSelector] before a string or string array to get a selection list in the inspector.

If used before an array, a list will appear for every array value.

Included a test file (TagSelectorTestScript).
Drag it to any object to test.

Useful if we don't want to add tags using strings, in case we change tag names later.